FAR-51/refactor: 핫플레이스 리팩토링
