+++ src/main/java/com/ssafy/faraway/domain/hotplace/service/impl/HotPlaceQueryServiceImpl.java
@@ -1,11 +1,13 @@
-package com.ssafy.faraway.domain.hotplace.service;
+package com.ssafy.faraway.domain.hotplace.service.impl;
 
 import com.ssafy.faraway.domain.hotplace.dto.req.HotPlaceSearchCondition;
 import com.ssafy.faraway.domain.hotplace.dto.res.HotPlaceCommentResponse;
+import com.ssafy.faraway.domain.hotplace.dto.res.HotPlaceImageResponse;
 import com.ssafy.faraway.domain.hotplace.dto.res.HotPlaceResponse;
 import com.ssafy.faraway.domain.hotplace.dto.res.ListHotPlaceResponse;
 import com.ssafy.faraway.domain.hotplace.entity.HotPlace;
 import com.ssafy.faraway.domain.hotplace.repository.HotPlaceQueryRepository;
+import com.ssafy.faraway.domain.hotplace.service.HotPlaceQueryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -24,6 +26,7 @@ public class HotPlaceQueryServiceImpl implements HotPlaceQueryService {
     public HotPlaceResponse searchById(Long hotPlaceId) {
         HotPlace hotPlace = hotPlaceQueryRepository.searchById(hotPlaceId);
         List<HotPlaceCommentResponse> commentResponses = getCommentResponses(hotPlace);
+    List<HotPlaceImageResponse> imageResponses = getImageResponses(hotPlace);
         hotPlace.updateHit();
         return HotPlaceResponse.builder()
                 .id(hotPlace.getId())
@@ -37,7 +40,16 @@ public class HotPlaceQueryServiceImpl implements HotPlaceQueryService {
                 .rating(hotPlace.getRating())
                 .createdDate(hotPlace.getCreatedDate())
                 .commentResponses(commentResponses)
+                .imageResponses(imageResponses)
                 .build();
+    }
+
+    private List<HotPlaceImageResponse> getImageResponses(HotPlace hotPlace) {
+        return hotPlace.getHotPlaceImages().stream().map(hotPlaceImage -> HotPlaceImageResponse.builder()
+                .id(hotPlaceImage.getId())
+                .fileName(hotPlaceImage.getUploadFile().getUploadFileName())
+                .createdDate(hotPlaceImage.getCreatedDate())
+                .build()).collect(Collectors.toList());
     }
 
     private List<HotPlaceCommentResponse> getCommentResponses(HotPlace hotPlace) {
